auth/requirements.txt
astroid==3.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
Flask==3.0.2
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
mariadb==1.1.10
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
parso==0.8.3
platformdirs==4.2.0
PyJWT==2.8.0
pylint==3.1.0
tomlkit==0.12.4
Werkzeug==3.0.1
